Switch our source of truth of source code dependencies to one based on cquery.
Note that this doesn't change how we build alchemist, just the decision
of whether we need a rebuild.

The current mechanism is very prone to missing files, as it relies on a
hardcoded list of crates we depend on. If we were to add extra
dependencies, it's likely they'd get missed.

BUG=None
TEST=portage/tools/run_tests.sh

Change-Id: Icd4d59c63422d844ed05a3b83845d84762026a8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5034084
Commit-Queue: Tim Bain <tbain@google.com>
Auto-Submit: Matt Stark <msta@google.com>
Tested-by: Tim Bain <tbain@google.com>
Reviewed-by: Tim Bain <tbain@google.com>