Enable --experimental_remote_cache_async.
This flag moves the "uploading outputs" step out of the action and into
an async thread. This enables dependent actions to start executing
immediately without being blocked on network activity (assuming there's
a free core).

In a full build, there are ~594 such "uploading outputs" tasks, each
taking anywhere from sub-second to more than a minute.

I suspect that given the long time chromeos-base/chromeos-chrome takes
by itself in the critical path in the latter of the build, any deferred
uploads can happen in parallel then, therefore speeding up the former
half of the build.

https://ui.perfetto.dev/#!/?s=6dfcd3e7b11dd64f5f25004048a20343cb62edcf1ce8bf35cabbebc05fb2d83e
-> search "upload outputs"

Note that this is experimental, and we should pay close attention to the
e2e latency of postsubmit builds once this is submitted.

BUG=b:311319090
TEST=CQ

Change-Id: I5433a93c74dd79bb608b815b801dd2a95d0afe39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5035462
Tested-by: Jingwen Chen <jingwen@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Jingwen Chen <jingwen@google.com>