alchemist: Merge ThinPackageRef to PackageRef
We used to have ThinPackageRef, a subset of PackageRef meant to be used
when performing atom matches before package metadata are fully computed.
However it turned out that this design is not very useful because we
even need the middle of the two; for example, on computing USE flags of
a package, we already know SLOT of the package but USE flags themselves
are not yet available, thus neither PackageRef and ThinPackageRef are
suitable to represent the available information of packages.

This patch removes ThinPackageRef, and instead modify PackageRef to make
some fields optional. Some fields, namely package_name and version, are
still mandatory because they should be always available (aka BasicInfo).
Predicates should ignore missing fields.

BUG=b:303400631
TEST=bazel test //bazel/portage/bin/alchemist:all

Change-Id: I1c88477e80f13203e47a03a9496c8f6987f0e700
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5040072
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>