README: Add some documentation regarding the different build stages
This describes the architecture on how BDEPENDs come into play, and what
all the various stages are for. I decided to put this in our main README
for now. We can migrate it to the chromiumos docs when we migrate all
the bazel docs over.

BUG=b:311670091
TEST=mdformat

Change-Id: Id64cd603fd8933b71cbc3959ca5451cb082f763c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5042055
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Tim Bain <tbain@google.com>