alchemist: Decouple cross_compile from DependencyKind
When DependencyKind is defined to represent a package dependency type,
associating "whether the package is being emerged in cross-root" to it
is weird because it's not a static property of a dependency. In fact,
the field is used only to decide whether to apply hard-coded
dependencies to packages.

This patch decouples the field from DependencyKind, and instead pass it
along to get_extra_dependencies as a function argument.

This is a pure cleanup to prepare to use DependencyKind as a key to the
fields in PackageDependencies.

BUG=b:299554821
TEST=bazel test //bazel/portage/bin/alchemist:all

Change-Id: I189a9da95333c0e3f4e3e7ebfe16da60bf7d4dc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5075544
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>