alchemist: Generate @portage//internal/bashrcs
When compiled profiles are enabled, we get the bashrc files from the
overlay layer. This change introduces a new `internal/bashrcs` directory
that we can use instead of relying on the overlay layer. I chose this
approach instead of generating new BUILD.bazel files in the overlay
directories because it would break the existing globbing we rely on and
generating different targets that accumulate their children would
add a lot more complexity.

I'm using the name of the bashrc file as the target name to avoid
introducing a conversion function.

The ebuild target will be updated to consume this new provider.

BUG=b:262460688
TEST=cargo test
TEST=BOARD=amd64-generic bazel build @portage//internal/bashrcs/...

Change-Id: I349cb5a051abd6f7a8f7341301ce0d00539f29f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5068788
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>