portage: Exclude packages file from overlays
`packages` is used to define the implicit system set. It's an analysis
file, so we don't need it.

I want to get rid of the `categories` file as well, but we would need to
generate the file for the specific package we are building, otherwise
the `emerge-$BOARD` command will fail (not that anyone should be using
it to begin with).

```
$ emerge-amd64-host -pv sys-libs/glibc

These are the packages that would be merged, in order:

Calculating dependencies /
!!! 'sys-libs/glibc' has a category that is not listed in /build/amd64-host/etc/portage/categories
... done in 0:00:01.040876

emerge: there are no ebuilds to satisfy "sys-libs/glibc" for /build/amd64-host/.
```

BUG=b:262460688
TEST=BOARD=amd64-generic bazel build --//bazel/portage:enable-portage-config-validator@portage//internal/packages/stage1/target/host/portage-stable/app-shells/bash:5.1_p16-r2

Change-Id: Ief4bd908b7ad038bf59642c4a498d747a02bec56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5068790
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>