Split documentation into multiple pages
The current documentation mixes contents for users and developers. Split
them into multiple pages to make it easier to read for end-users.

This patch mostly keeps existing contents, but mostly just relocates
them.

Note that eventually we may want to move user-facing docs to the new
still-under-construction central repo (go/cros-dev-library) or the
deprecated central repo (go/chromiumos-docs) (which are an ideal example
of https://goomics.net/50/), but let us keep our docs in our repos for
now for faster iterations. It should be trivial to move to new repos
anyway as long as we write them in Markdown.

BUG=b:313535588
TEST=https://chromium.googlesource.com/chromiumos/bazel/+/refs/changes/14/5076014/5/README.md

Change-Id: Ib6609692779b2f9afbc03a1925001ffd51bb6704
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5076014
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Tim Bain <tbain@google.com>
Tested-by: Shuhei Takahashi <nya@chromium.org>