alchemist: Make RepositoryLookup return a "partial" RepositorySet
RepositoryLookup is a ChromeOS specific method that's only used when
building a "fakechroot". Reusing the RepositorySet which is a portage
concept resulted a flawed `primary()` method getting added to it.

This change makes RepositoryLookup return a "partial" repository set
that can be used to generate the make.confs for the fake chroot, while
also tracking the "primary" or "root" repository that was used to lookup
all the other repos.

BUG=b:262460688
TEST=cargo test

Change-Id: I1d8967d089aee3a692225ba2965485986f73c4e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5078254
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>