alchemist: Add a name to RepositorySet
Right now we are using the name of the "primary" repository as the name
of the repository set. This is confusing because RepositorySet.load
actually takes in a portage config root, not a specific overlay root.
The config root can have a PORTDIR_OVERLAY that contains a hand curated
list of repositories. So it doesn't really make sense to use the name of
the "primary" repository. The next CL will change the generators to use
the name().

The API for RepositoryLookup.create_repository_set is a bit strange now
that we introduced the name. I'm using the name of the "board" as the
name of this repository set. Ideally we wouldn't be returning a
RepositorySet since the order computed by traversing the parents isn't
exactly the priority of the repositories. I'll clean this up in a follow
up.

BUG=b:262460688
TEST=cargo test

Change-Id: I509f79629f7a08f4b425dd9f89fbe23a0d2d5848
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5078249
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>