alchemist: Replace chromeos/config/chromeos_version.sh with a stub
chromiumos-overlay/chromeos/config/chromeos_version.sh is updated daily,
so it should not be included in the overlay targets as it would trigger
rebuilding all packages every day. OTOH, it's prone to wrong build
results if we simply remove it from the ephemeral environment as many of
the script's callers check its existence and skip calling it when it is
missing.

This patch replaces the script in the overlay target with a stub that
immediately terminates the ephemeral container when it is called. This
way, we can detect a package that calls the script without properly
delcaring extra source dependencies.

BUG=b:314873346
TEST=BOARD=amd64-generic bazel build \
     @portage//target/virtual/target-os{,-dev,-test}:package_set
TEST=bazel test //bazel/portage/bin/alchemist:all

Cq-Depend: chromium:5330787
Change-Id: Ib7e93db7bc4adfac24cdbac5d30a94aadfa769a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5087055
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>