alchemist: Add profile_path() to TargetData
We were losing the original profile path. This change exposes the
original profile path.

When running inside the chroot, the --profile flag currently determines
the profile symlink we use when building packages. This results in a
mismatch between the real profile we use for analysis that was defined
by /etc/portage/make.profile and the --profile flag we use for building
packages. By keeping the original profile path, we can start to clean up
this inconsistency.

BUG=b:262460688
TEST=cargo test

Change-Id: I1658156425c84e8dd948a0c412d859ba3d48b93d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5078252
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>