alchemist: Override PKGDIR, PORTAGE_TMPDIR, etc when running in chroot
When running inside the chroot we use the actual configs generated by
setup_board. When the amd64-host board gets generated the `PKGDIR`, etc
variables are still pointing to the BROOT. In theory we could update
the setup_board command, but it gets tricky because we don't normally
cross-root compile this board in the portage flow. So we would need to
ensure that this config didn't persist into the SDK tarball.

Why is this a problem now? With compiled profiles we use the config
bundle to write out a generated make.conf. This means we need to
variables to be correctly set. This isn't a problem when running outside
the chroot, because we properly set these variables in
`generate_make_conf_host_setup`.

BUG=b:262460688
TEST=(chroot) BOARD=amd64-generic /mnt/host/depot_tools/bazel build @portage//target/chromeos-base/metrics

Change-Id: I1bf8883116b5ddc89c95f96100c278fc87ec7b7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5113119
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>