alchemist: Remove all_details argument from analyze_packages
analyze_packages used to take a list of packages (as
Vec<MaybePackageDetails>) specifying which packages to analyze.
However, in reality, it's not useful to pass a list smaller than the
complete list of targets because install set computation fails if
a package directly or indirectly depends on a package outside the
original package set.

Thus this change moves the PackageResolver::find_all_packages call
from load_packages to analyze_packages so that we always analyze the
whole set of packages.

Alternatively, we could leave analyze_packages to still take the
packages argument and lazily analyze packages by introducing a cache
(like PackageResolver), which would allow dump_package subcommand to
print transitive dependencies. I'm not doing this way just because it's
going to be a large change while the only use case is dump_package for
now.

BUG=b:317310816
TEST=bazel test //bazel/portage/bin/alchemist:all

Change-Id: I18f36bdb8f973a4ce32d32b99d5da68429837e3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5149439
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>