alchemist: Ignore PDEPEND on computing build_host_deps (again)
crrev.com/c/5144089 changed the dependency analyzer to consider PDEPEND
on computing build_host_deps, but it turned out unnecessary; PDEPEND is
effective only on computing the set of packages to eventually install on
the user system, not during building packages.

Also updates the comment of build_host_deps to clarify what it contains.
Particuarly an important clarification is that it doesn't contain all
transitive dependencies. An unit test is added to ensure the behavior.

BUG=b:317310816
TEST=bazel test //bazel/portage/bin/alchemist:all

Change-Id: I3f36e69e34036388869bb25002255613f61f6d19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5154727
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>