run_in_container: Switch into the new root before pivot_root
I was trying to debug why run_in_container was printing out a file
not found error when building a Stage 3 SDK. As part of debugging I
sprinkled some `std::env::current_dir` and was surprised to see that
it failed after the `umount /host`. This change makes it so that the
CWD gets updated to point to "/" after the pivot_root. This ensures we
always have a valid CWD.

The problem I was debugging was not fixed by this, it turns out my
stage 3 sdk was missing bash, but I thought it would be a nice clean up.

BUG=none
TEST=BOARD=amd64-generic bazel build @portage//internal/sdk/stage4

Change-Id: If74a8aadacb903669b84972386e129fbda0b30f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5179193
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>