portage: Clear ebuild file after package installation
We don't need the ebuild file after a package has been installed. By
clearing it we can avoid cache busting all the ebuild targets that
depend on it when the ebuild is changed. We will still cache bust the
`_deps` layers, but those are quick to recreate.

This change is helpful when hacking on an ebuild locally. It will also
be helpful when we switch to 9999 ebuilds.

BUG=b:317120571
TEST=BOARD=amd64-generic bazel build @portage//internal/sdk/stage4

Change-Id: I1ed49b0a13040f0140ccf8e7b32f6944748f0f5b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5179093
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>