alchemist: Generate Stage 3 Bootstrap and Stage 4 SDK targets
This will allow us to build a Stage 3 Bootstrap SDK that can be used
as a Stage 1 tarball. The generated SDK is built by finding the
transitive DEPEND and RDEPEND of the implicit system package and
collecting their BDEPENDs. This ensures that the bootstrap SDK has all
the necessary host tools available to cross-root compile the implicit
system.

The Stage 4 target is really a short hand for:
    bazel build @portage//internal/packages/stage3/target/host/chromiumos/virtual/target-sdk-implicit-system:1-r4

BUG=b:319154851
TEST=BOARD=amd64-generic bazel build @portage//internal/sdk/stage4

Change-Id: Ib51d7ba2ac2a859934a463c577118e25f6bf1a4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5181792
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>