portage: Define extra_sources rule
This macro essentially wraps rules_pkg to generate a tarball, but it
provides a safe subset of rules_pkg's functionality so that users are
less likely to misuse them. Particularly, a directory prefix (i.e.
/mnt/host/source/src/<package-path>) is automatically derived from the
current package, it's forbidden to rename files, and this is the only
way to define valid targets for extra sources as it creates a private
provider ExtraSourcesInfo.

BUG=b:321628179
TEST=With later CLs

Change-Id: Ie3000202576c9a55a54356327af064bffe4b3780
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5225126
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>