alchemist: Replace __xbuild with __alchemist
All internal variables in ebuild_prelude.sh used to be prefixed with
__xbuild. "xbuild" (cross-build) was the name of my exploratory project
to replace the CrOS build system that I started before Alchemy, and the
name was still kept in ebuild_prelude.sh just because I simply copied
the file from the previous project. Now that we've got a fancy project
name, it's time to drop the old name.

BUG=None
TEST=bazel test //bazel/portage/bin/alchemist:all

Change-Id: Id40e26e924c22ad0e3c70d1161d9c92364bd3120
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5232960
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>