bazel: Replace @portage//:all_packages w/ @portage//:__subpackages__
This will remove the @portage repository from the transitive deps of
all our host tools. This means we no longer need to re-validate the
Stage 1 SDK's extracted contents when we make a portage change. This
results in a ~50 second reduction in the ebuild edit-build cycle.

BUG=b:299131906
TEST=BOARD=amd64-generic bazel build @portage//internal/sdk/stage2/target/board

Cq-Depend: chromium:5273925
Change-Id: Ic9edcb01f130441aed458feff1659c9f009bb9ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5273094
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Tim Bain <tbain@google.com>
Commit-Queue: Raul Rangel <rrangel@chromium.org>