portage: Different message for EbuildValidation
Specify a different progress_message for the EbuildValidation target, so
that this action's execution can be distinguished from the Ebuild.

This is especially useful for analyzing the JSON trace profiles, because
they use progress_message as the name by default, and other properties,
like the mnemonic or target path, are harder to notice or (for critical
path items) not displayed at all.

BUG=b:291987893
TEST=none

Change-Id: I2b06c051c1e199897c9c09fc9fa664fdeee7bdb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5272903
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Auto-Submit: Maksim Ivanov <emaxx@chromium.org>