alchemist: Make IDEPEND honor EAPI
I had originally let the IDEPEND variable leak through regardless of the
EAPI since IDEPEND is required when building under bazel, but it
is causing problems. This change makes it so we ignore IDEPEND if the
EAPI doesn't support it. If we need to add IDEPENDs for a package, we
will need to do it in our hack file.

BUG=b:303982248
TEST=BOARD=amd64-generic bazel build -@portage//target/virtual/target-os:package_set @portage//target/virtual/target-os-dev:package_set @portage//target/virtual/target-os-factory:package_set @portage//target/virtual/target-os-factory-shim:package_set @portage//target/virtual/target-os-test:package_set @portage//target/chromeos-base/autotest-all:package_set

Force-Relevant-Builds: amd64-generic-bazel-cq
Change-Id: Ib35cf119a06323edd7790920baded47fff7b385c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5279015
Reviewed-by: Tim Bain <tbain@google.com>
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>