fakefs: Rewrite testhelper in C
testhelper is currently written in Go. This means that the produced
binary is statically linked without any dependency to libc. This turned
out to be inconvenient when we want to use the binary to test
libfakefs_preload.so.

This patch rewrites testhelper in C so that it's dynamically linked to
libc. Fortunately, the logic there is simple so it's not too annoying to
write it in pure C.

Also adds a small unit test that ensures that testhelper is a
dynamically linked binary.

BUG=b:324515572
TEST=bazel test //bazel/portage/bin/fakefs/...

Change-Id: Ic54127017afc76a9eb7e58fb8a93a4a23dd15466
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5279020
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>