ContainerSettings: use mutable_base_dir for archive expansion tmpdir
The archive expansion tmpdir currently defaults to being created under
std::env::temp_dir(), which is not a suitable location when running in
a docker container (/tmp points to an overlayfs filesystem, which can
lead to attempts at stacking 3 layers high of overlayfs). This comes
up when trying to build packages on RBE.

Using mutable_base_dir as a base directory under which to create the
archive expansion temp dir is in line with the rest of temp dir
creations in ContainerSettings, and allows us to specify a base that
is not overlayfs mounted.

BUG=b:323923147
TEST=BOARD=amd64-generic bazel build \
@portage//target/virtual/target-os:package_set
TEST="locally" tested that this resolves an issue with RBE builds of
packages with over 500 overlayfs lower layers

Change-Id: Ibba247ce1a5295528e33072210359025d35195c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5297415
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Tested-by: Romain Jobredeaux <jobredeaux@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Romain Jobredeaux <jobredeaux@google.com>