alchemist: Fix incorrect glibc package.provided check
An atom is considered provided if it matches a package in
package.provided. The logic in `get_primordial_packages` was flawed
because it was expecting the found package version to match the provided
package. This change fixes this by switching over to using
`find_provided_packages`.

BUG=b:266979761
TEST=cargo test

Change-Id: I1904d21eb09320b982b841911af6c17a803863f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5302635
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>