Decompress environment.bz2 in advance for staged contents layers
Portage compresses the environment file in XPAK/VDB with bzip2, but
the cost to compress/decompress bzip2 archives is not negligible when
fast package installation is keen on performance. Also, it needs
bzip2 CLI installed on the system, which we want to avoid.

This patch updates extract_package to decompress environment.bz2 to
environment.raw when generating staged package contents layers.
drive_binary_package now reads the environment from the raw file instead
of the bz2 file.

BUG=b:313557016
TEST=bazel test //bazel/portage/bin/...

Change-Id: I2adf9a59bbd750e5dcaf2d06a05c586ea130b8b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5330731
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>