Introduce auditfuse
auditfuse is a FUSE file system that audits file accesses on a
read-only view of another directory. This is used to check if package
hooks access the file system.

I wrote this daemon in Go because hanwen/go-fuse is the best FUSE
library I've ever seen, providing high-level API on top of the raw FUSE
interface.

BUG=b:313557016
TEST=bazel test //bazel/portage/bin/auditfuse/...

Change-Id: I21b17165e998cd1da0c6a5343a63c25524b23891
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5330733
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>