extract_package: Audit package hooks
When extract_package creates a staged contents layer, it will attempts
to run hooks with auditfuse to see if they access any file on the file
system, and saves the result to CROS_BAZEL_HOOK_REQUIRES in the VDB dir.
In a later patch, fast_install_packages will read the file to see if it
needs to run hooks.

BUG=b:313557016
TEST=bazel test //bazel/portage/bin/extract_package:all

Change-Id: I818d0c04ec74d6d3a92f8f339f77356053397a80
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5330734
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>