alchemist: Compute *DEPEND XPAK keys for each package
We are stripping sub-slots and revision numbers from the ephemeral
chroot. We are doing this to reduce the number of targets that get
invalidated on an ebuild only change, and to enable interface libraries.
Since the ephemeral chroot no longer contains the real sub-slot the
*DEPEND keys that get baked into the resulting binpkg are incorrect.
This means we need to post-processes the binpkg to add this information
back in.

This change adds the logic to compute the portage equivalent *DEPEND
XPAK entries.

You can find the original portage logic here:
https://chromium.googlesource.com/chromiumos/third_party/portage_tool/+/refs/heads/chromeos-2.3.75/lib/portage/dep/_slot_operator.py#52

BUG=b:311094915
TEST=BOARD=amd64-generic bazel build @portage//target/chromeos-base/vboot_reference:installed
TEST=cargo test

Change-Id: I9078c8f1e6287f541b2f0421bd31ca70e402df81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5340603
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>