alchemist: Run end-to-end golden tests in a subprocess
We used to run end-to-end golden tests of alchemist in a unit test. But
this approach came with challenges: multi-threaded unit tests were
difficult to work with user/mount namespaces used to generate fake
chroot. We had some hacks to mitigate the issues, such as a flag to skip
calling unshare(2), but it was difficult to avoid environment skews when
running tests inside/outside chroot.

Also we had an issue where alchemist unit test binary calls itself
recursively when an ebuild calls ver_rs or ver_test (b/328710965).

This patch gives a fundamental solution to these problems by running
alchemist as a subprocess of the unit test binary. Also, to minimize the
environment difference when tests are run inside/outside the chroot,
introduce a few flags to alchemist that control its behavior that were
formerly controlled by auto-detection of the chroot.

BUG=b:328710965
TEST=bazel test //bazel/portage/bin/alchemist/... # inside chroot
TEST=bazel test //bazel/portage/bin/alchemist/... # outside chroot

Change-Id: I15f6a912b84890dbf3079b7b8c4d43e45f9e2648
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5353333
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>