portage: Add option to return interface layers to sdk_install_deps
This change adds the `contents` attribute to
`sdk_install_deps`. It allows choosing the type of content layers
we want to return from `sdk_install_deps`.

I don't currently plan on supporting `SDEPEND`, so I don't need a knob
to select which deps we install as interface layers, and which ones we
install as sparse layers. I plan on just having a
`supports_interface_libraries` bool that can be set in the package
specific toml file.

BUG=b:261863720
TEST=BOARD=amd64-generic bazel build @portage//target/dev-libs/libgcrypt

Change-Id: I38108f75878ad628856d3c667503f112bd6dfd70
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5370141
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>