ebuild: Always rerun :installed targets
We always want the :installed targets to run because the portage sysroot
has a different lifetime than the bazel execroot. In order to do this
we need to leverage the CACHE_BUST_DATE so we can create a file that
always changes. We then make this an input to the :installed action.

BUG=b:328077442
TEST=BOARD=brya bazel build @portage//target/chromeos-base/vboot_reference:installed

Change-Id: I146f8b52570d95848a9576d2b5ed6d276453fdbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5386329
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Tim Bain <tbain@google.com>
Commit-Queue: Raul Rangel <rrangel@chromium.org>