process_artifacts: Handle deeply nested named sets efficiently
Named sets in BEP can be deeply nested with duplication, which resulted
in process_artifacts eating up all RAMs on bots with the previous naive
implementation.

This patch introduces a data structure that provides fast merge
operations over named sets. The key is to keep track of named set IDs
that are already merged to the set, not only a set of files, to avoid
repeatedly merging the same named sets.

BUG=b:297602342
TEST=bazel test //bazel/portage/tools/process_artifacts:all

Change-Id: Ic37e631bfd1d30778cb27f4bba73241a9227d4d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5404093
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>