action_wrapper: Record rusage in logs and profile JSON
This patch adds getrusage call to get the CPU time spent by the action,
and prints it to logs, as well as recording it in the profile JSON.

The profile JSON will get a new process for action_wrapper, whose PID is
always set to 1. It will have the only span containing resource usage
data.

BUG=None
TEST=bazel build ... # and inspect *.profile.json

Change-Id: I0c8ffb099a11f621f9dd2a48cc0b35210d8c40d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5407215
Reviewed-by: Tim Bain <tbain@google.com>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>