alchemist: Remove ACCEPT_LICENSE from ebuild environment
This variable is a space separated list of licenses and license groups.
It needs to be analyzed at generate-repo time, not when the ebuild is
invoked. Strip it out of the environment so that we can get cache hits
between CI builders and local builds.

BUG=b:332588234, b:309622826
TEST=cargo test

Change-Id: I27fec283c839922dbdea139c5cf13c1fdd0bb888
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5417233
Tested-by: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>