Adds a Bazel flag to replace the stage1 SDK tarball with a fake
This patch introduces a Bazel flag
--//bazel/portage/sdk:omit_stage1_tarball that replaces the stage1 SDK
tarball with a fake one. This allows us to run `bazel cquery` without
downloading the huge stage1 SDK tarball.

This flag is very similar to --//bazel/portage:omit_ebuild_src. Since
it's enabled by default on `bazel cquery` in common.bazelrc, this patch
enables the new flag by default as well. Honestly I don't like it very
much to mutate the dependency graph implicitly on cquery, but it is what
it is as of today.

BUG=b:333002601
TEST=BOARD=amd64-generic bazel cquery \
     'deps(@portage//target/sys-apps/attr)' | sort | \
     grep //bazel/portage/sdk
TEST=bazel build //bazel/...

Change-Id: Ifad8667c55e446ae774f2acc3546a4591a6eddca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5442193
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>