Seperate cargo from bazel completely.
Previously, we had a dependency loop:
* The rust crates module extension depends on bazel/Cargo.toml
* bazel/Cargo.toml depends on bazel/portage/bin/build_package/Cargo.toml
* bazel/portage/bin/build_package/Cargo.toml depends on //bazel/portage/bin/build_package:build_package
* //bazel/portage/bin/build_package:build_package depends on the rust crates module extension.

This dependency loop was a non-issue most of the time, but it becomes very difficult to work with at times.

To solve this, we simply make alchemy_crates/Cargo.(toml|lock) the source of truth, and make everything else autogenerated.

BUG=b:333304142
TEST=portage/tools/run_tests.sh

Change-Id: I4a98b4b639cec0c89569df3b3e1c03deecd7688d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5449315
Commit-Queue: Matt Stark <msta@google.com>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Matt Stark <msta@google.com>