process_artifacts: Add a --prebuilts option to generate prebuilts.bzl
This action is similar to archive_logs, except it generates a a list of
all the binpkgs produced/consumed by the build and their CAS locations.

This is similar to the `generate-stage2-prebuilts` script, but not as
hacky. I chose to consume the ebuild metadata instead of the binpkgs
themselves because in a cached build we should never need to download
the stage1 packages, and I didn't want to force them to download just
so I could extract CAS location from the BES log.

I plan on having the CI builders publish the prebuilts.bzl. This will
make development of primordial targets easier since we can easily pin
the packages we aren't interested in. In the future it might be nice
to include the manifest SHA that was used to perform the build, as well
as the BOARD. Bazel doesn't know about those two things so they aren't
part of the BES.

BUG=b:331947109
TEST=cargo test

Change-Id: I772da5c296a45eb85d723c24f46e99b034506024
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5473594
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>