portage: Add enable_hermetic_validator flag
We currently have `_hermetic_test` targets that can be ran with bazel
test. These are convenient when iterating on a single package, but
difficult to use when trying to test all transitive deps. As a
workaround we had _package_set_hermetic_test, but it only coverts the
install_set of a package, it doesn't cover all the build time
dependencies or build tools. This change introduces a new flag that when
set, it will add a validator to each ebuild action that will compare the
package with the _hermetic_test_package. This will make it easy to run
hermetic tests for all packages in the graph.

BUG=b:266973461
TEST=BOARD=brya bazel build --//bazel/portage:enable_hermetic_validator  @portage//target/sys-firmware/brya-kernel-ucode

Change-Id: Idb732d86997666527c2838429c57e841da049fa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5490762
Tested-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Auto-Submit: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Tim Bain <tbain@google.com>
Commit-Queue: Tim Bain <tbain@google.com>