extract_package: Always enable verbose outputs of auditfuse
Performance doesn't matter for auditfuse, so let's always enable
verbose outputs for better debuggability. We still keep the FUSE
protocol debugging off by default since it's too noisy.

The main part of this patch is actually to support --verbose and --debug
in the daemon mode (no --foreground). Formally, in the process of
daemonization, the daemon process communicated with the first process
via stderr, but this patch changes it to use FD 3 so that the daemon
process keep the ability to write messages to stderr.

BUG=b:338093333
TEST=BOARD=amd64-generic bazel build @portage//internal/packages/\
     stage1/target/host/chromiumos/virtual/baselayout-debug:1

Change-Id: I6e7f1d7a1b61f27da433b1287f2e476692edfef6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5513860
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>