portage: Add ContentLayerTypesInfo to represent full and internal layers
With build_sdk switching over to using fast_package_install, we lost
the full vdb in the generated SDK layer. This was intentional since we
don't want to cause full rebuilds when this metadata changes. However
when generating the SDK tarball, we want to keep the full vdb metadata.
This will allow the tarball to fully function with `emerge` and
`cros_sdk`.

This change introduces two new content layers. These layers always have
a full vdb, and also a full revision number. We can pass in the full_vdb
argument to `sdk_install_deps` to generate an SDKInfo with the full
layers.

Here is a diff comparing the SDK tarball built using `emerge` and the
SDK tarball built using fast_package_install.

https://paste.googleplex.com/4774984501428224#l=81

We still have a few diffs to sort out.

Having 4 different layers of duplicated contents is expensive, but we
only need the `full` layers when building an SDK or an OS image (using
bazel). Ideally we could introduce a common "content layer" that
contains only the extracted files. We could then have a full and sparse
vdb layer. This will reduce the number of times we extract, but it does
increase the number of layers.

BUG=b:308661552
TEST=BOARD=brya bazel build --config=prebuilts  @portage//internal/sdk/stage2:tarball

Change-Id: If813ebb099c52b8f927a9be7754e4b5b94fc1608
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/bazel/+/5522986
Commit-Queue: Raul Rangel <rrangel@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Tested-by: Raul Rangel <rrangel@chromium.org>